+++ src/extension.ts
@@ -12,6 +12,7 @@ export const CONFIG_CLEAN_START = "cleanStart";
 export const CONFIG_REGEX = "regex";
 export const CONFIG_COLOR = "color";
 export const CONFIG_BACKUP = "backup";
+export const CONFIG_COMBINE_TOGGLE = "combineToggle";
 
 export const CONFIG_TOKENS = "tokenColorCustomizations";